ref: move shared highlighting keys into a core leaf
PhelAnnotationConstants and PhelTextAttributesRegistry are TextAttributesKey
holders with no behaviour, and each key's external id is persisted in the user's
colour scheme — they are contract, not internals. Living under annotator's and
syntax's subfolders left the colorsettings page, whose job is to enumerate every
key in the plugin, importing past two feature packages' roots to reach them.

Move both into core/highlighting: a leaf every feature may depend on, the same
shape as registry. That drops the last cross-feature internals imports; the
remaining editor -> syntax.PhelSyntaxHighlighter edge is a root import the
ColorSettingsPage API requires, and stays.

The external key strings are literals inside createTextAttributesKey and are
unaffected by the package move, so existing user colour schemes are preserved.
Unit tests follow the sources into unit/core/highlighting/.
