+++ src/test/kotlin/org/phellang/integration/annotator/PhelDeprecatedAnnotatorTest.kt
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
-import org.phellang.annotator.infrastructure.PhelAnnotationConstants
+import org.phellang.core.highlighting.PhelAnnotationConstants
 import org.phellang.registry.PhelFunctionRegistry
 
 class PhelDeprecatedAnnotatorTest {
