+++ src/test/kotlin/org/phellang/integration/syntax/PhelSyntaxHighlightingWorkflowTest.kt
@@ -8,7 +8,7 @@ import org.junit.jupiter.params.provider.CsvSource
 import org.phellang.syntax.PhelSyntaxHighlighter
 import org.phellang.language.psi.PhelTypes
 import org.phellang.language.lexer.PhelLexerAdapter
-import org.phellang.syntax.attributes.PhelTextAttributesRegistry
+import org.phellang.core.highlighting.PhelTextAttributesRegistry
 import org.phellang.syntax.mapping.PhelTokenAttributeMapper
 import org.phellang.syntax.classification.PhelTokenClassifier
